requirements.txt
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.1.1
certifi==2026.2.25
colorama==0.4.6
distro==1.9.0
fastapi==0.104.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jinja2==3.1.2
jiter==0.13.0
openai==2.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.12.5
pydantic_core==2.41.5
pymongo==4.6.1
python-dotenv==1.2.2
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uv==0.10.8
uvicorn==0.24.0
